store-backend/pyproject.toml
[project]
name = "store-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "roger.barriga",email = "roger.barriga@scribsoft.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.12.3,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi (>=0.120.4,<0.121.0)",
]

[tool.poetry]
packages = [{include = "store_backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
